grid patched terrain

+++ Files/Headers/ObjectManager.h
@@ -17,9 +17,9 @@ namespace SweepingBirds
 		void createAILogger();
 
 		bool bind_object(Textured3DObject* object, int nb_instances, int index);
-		const std::vector<std::pair<Textured3DObject*, float*> >& get_objects() { return m_vObjectManaged; };
-		std::pair<Textured3DObject *, float *>& get_object(int index);
-		GUIInfos * generate_slider_nb_instances_infos(int index, float max);
+		const std::vector<std::pair<Textured3DObject*, int*> >& get_objects() { return m_vObjectManaged; };
+		std::pair<Textured3DObject *, int *>& get_object(int index);
+		GUIInfos * generate_slider_nb_instances_infos(int index, int max);
 		GUIInfos * generate_slider_cube_size(int index);
 		GUIInfos * generate_slider_cube_radius_spacing(int index);
 		GUIInfos * generate_slider_cube_speed(int index);
@@ -28,7 +28,7 @@ namespace SweepingBirds
 		GUIInfos * generate_slider(std::string name, float min, float max, float step, float * value_to_change);
 	private:
 		int m_iNbObjectManaged;
-		std::vector<std::pair<Textured3DObject*, float*> > m_vObjectManaged;
+		std::vector<std::pair<Textured3DObject*, int*> > m_vObjectManaged;
 
 	};
