+++ Files/Headers/SceneManager.h
@@ -38,7 +38,7 @@ namespace SweepingBirds
 		bool setup_skybox();
 
 		void draw_scene(ShaderProgram * shader, glm::mat4 projection, glm::mat4 wtv);
-		void draw_object(std::pair<Textured3DObject *, float*> object, ShaderProgram * shader, glm::mat4 projection, glm::mat4 wtv);
+		void draw_object(std::pair<Textured3DObject *, int*> object, ShaderProgram * shader, glm::mat4 projection, glm::mat4 wtv);
 		void gamma_management(glm::mat4 mvp, glm::mat4 mv);
 		void debug_frame_buffer(glm::mat4 mvp, glm::mat4 mv, bool activate_shadow_map, bool activate_sobel_map);
